Remove duplicate posts from home page
Due to a breaking change introduced by Webpack 5 [1], `require.context` has started to include duplicate versions of files based on the possible import paths.

Next.js has upgraded to Webpack 5, which was enabled by default for version 11 [2]. Now, this project is on Next.js 12, which has deprecated Webpack 4 entirely.

This change includes the relative filepath prefix ("./") to the regular expression, so now it captures any files in the "posts" directory that starts with "./" and ends with ".md".

[1]: https://github.com/webpack/webpack/issues/12087
[2]: https://nextjs.org/docs/messages/webpack5